excluido componente sendBox e criado form no app, criado a lista de mensagens
