--- modulo1/projeto-whatslab/whatslab/src/components/SendMessage.css
@@ -1,27 +1,0 @@
-.inputBox {
-  display: flex;
-  width: 100%;
-  box-sizing: border-box;
-  display: flex;
-  font-family: serif;
-}
-
-.firstInput {
-  width: 30%;
-  padding: 10px;
-  border: 1px solid #ebebeb;
-  font-size: 16px;
-}
-
-.secondInput {
-  width: 60%; 
-  padding: 10px 44px 10px 10px;
-  border: 1px solid #ebebeb;
-  font-size: 16px;
-}
-
-.sendBtn {
-  width: 10%;
-  background-color: #fff;
-  border: 1px solid #ebebeb;
-}
